requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
jinja2>=3.1.4
aiofiles>=23.2.1
qrcode[pil]>=7.4.2
pillow>=10.3.0
ultralytics>=8.2.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
tensorflow>=2.16.0
scikit-learn>=1.4.0
